refactor(dpo): move chat-template render helpers to shared module for RL reuse
