+++ src/axolotl/prompt_strategies/dpo/chat_template.py
@@ -2,198 +2,14 @@
 DPO prompt strategies for using tokenizer chat templates.
 """
 
-import json
-
-from axolotl.prompt_strategies.jinja_template_analyzer import JinjaTemplateAnalyzer
-from axolotl.utils.chat_templates import extract_chat_template_args, get_chat_template
-from axolotl.utils.logging import get_logger
-from axolotl.utils.schemas.utils import handle_legacy_message_fields_logic
-
-LOG = get_logger(__name__)
-
-
-def _parse_tools(tools):
-    """Parse tools into a list of dicts, decoding JSON-encoded strings."""
-    if tools is None:
-        return None
-
-    if isinstance(tools, str):
-        try:
-            tools = json.loads(tools)
-        except json.JSONDecodeError as e:
-            LOG.error(f"Error parsing tools as JSON. Error: {e}")
-            raise
-
-    if isinstance(tools, list):
-        parsed_tools = []
-        for tool in tools:
-            # some datasets store each tool as a JSON-encoded string
-            if isinstance(tool, str):
-                try:
-                    tool = json.loads(tool)
-                except json.JSONDecodeError as e:
-                    LOG.error(f"Error parsing tool as JSON. Tool: {tool!r}, Error: {e}")
-                    raise
-            if isinstance(tool, dict) and "function" in tool:
-                function = tool["function"]
-                params = function.get("parameters")
-                if isinstance(params, str):
-                    try:
-                        function["parameters"] = json.loads(params)
-                    except json.JSONDecodeError as e:
-                        LOG.error(
-                            f"Error parsing tool parameters as JSON. "
-                            f"Function: {function.get('name', 'unknown')}, "
-                            f"Parameters string: {params!r}, "
-                            f"Error: {e}"
-                        )
-                        raise
-            parsed_tools.append(tool)
-        return parsed_tools
-
-    raise ValueError(
-        "Unknown tools format. Please convert it into a list[dict].\n"
-        f"Current format: {type(tools)}"
-    )
-
-
-def _parse_tool_call_arguments(message):
-    """Decode JSON-encoded tool call arguments so templates receive dicts."""
-    for tool_call in message.get("tool_calls") or []:
-        if "function" in tool_call and "arguments" in tool_call["function"]:
-            args = tool_call["function"]["arguments"]
-            if isinstance(args, str):
-                try:
-                    tool_call["function"]["arguments"] = json.loads(args)
-                except json.JSONDecodeError as e:
-                    LOG.error(
-                        f"Error parsing tool_calls arguments as JSON. "
-                        f"Function: {tool_call.get('function', {}).get('name', 'unknown')}, "
-                        f"Arguments string: {args!r}, "
-                        f"Error: {e}"
-                    )
-                    raise
-
-
-def _build_message_transform(message_property_mappings, role_map):
-    """Build a function that maps a raw dataset message to a chat template message,
-    preserving any extra properties the chat template uses (e.g. tool_calls)."""
-
-    def transform_message(message, msg_variables):
-        """Map a raw dataset message to a chat template message."""
-        transformed = {}
-        for target, source in message_property_mappings.items():
-            value = message.get(source)
-            if value is not None:
-                transformed[target] = value
-
-        if "role" in transformed:
-            transformed["role"] = role_map.get(transformed["role"], transformed["role"])
-
-        mapped_sources = set(message_property_mappings.values())
-        for key in msg_variables - mapped_sources:
-            value = message.get(key)
-            if value is not None:
-                transformed[key] = value
-
-        _parse_tool_call_arguments(transformed)
-        return transformed
-
-    return transform_message
-
-
-# always preserved: template analysis misses properties accessed via
-# `message.get(...)` (e.g. gemma4), so OpenAI message keys are unioned in
-_BASE_MSG_VARIABLES = frozenset(
-    ["tool_calls", "tool_call_id", "name", "reasoning_content", "reasoning"]
+from axolotl.prompt_strategies.chat_template_utils import (
+    build_message_transform,
+    make_msg_variables_getter,
+    parse_tools,
+    render_preference_sample,
 )
-
-
-def _make_msg_variables_getter():
-    """Cache chat template message variable analysis per template string."""
-    cache = {}
-
-    def get_msg_variables(chat_template_string):
-        """Return the message properties used by the chat template."""
-        if chat_template_string not in cache:
-            cache[chat_template_string] = (
-                JinjaTemplateAnalyzer(chat_template_string).get_message_vars("messages")
-                | _BASE_MSG_VARIABLES
-            )
-        return cache[chat_template_string]
-
-    return get_msg_variables
-
-
-DUMMY_USER_MESSAGE_CONTENT = "[[dummy_message]]"
-
-
-def _extract_response(full, prompt_prefix, content):
-    """Strip the rendered dummy-user prompt from a response rendering.
-
-    Strips the longest common prefix rather than requiring an exact prefix
-    match, since a generation prompt can diverge slightly from the completed
-    message rendering (e.g. thinking templates open `<think>` with different
-    whitespace than a rendered `reasoning_content` block).
-    """
-    common = 0
-    for prefix_char, full_char in zip(prompt_prefix, full, strict=False):
-        if prefix_char != full_char:
-            break
-        common += 1
-    response = full[common:]
-    if DUMMY_USER_MESSAGE_CONTENT not in response:
-        return response.rstrip()
-    # Fallback: locate the response content directly
-    if content:
-        strip_index = full.find(content)
-        if strip_index != -1:
-            return full[strip_index:].rstrip()
-    return full.rstrip()
-
-
-def _render_dpo_sample(
-    tokenizer,
-    messages,
-    chosen,
-    rejected,
-    chat_template_string,
-    chat_template_kwargs,
-    tools,
-):
-    """Render the prompt and extract the chosen/rejected response strings."""
-    template_kwargs = {
-        "chat_template": chat_template_string,
-        "tokenize": False,
-        **chat_template_kwargs,
-    }
-    if tools:
-        template_kwargs["tools"] = tools
-
-    dummy_user_message = {"role": "user", "content": DUMMY_USER_MESSAGE_CONTENT}
-
-    result = {}
-    result["prompt"] = tokenizer.apply_chat_template(
-        messages,
-        add_generation_prompt=True,
-        **template_kwargs,
-    )
-
-    dummy_prompt = tokenizer.apply_chat_template(
-        [dummy_user_message],
-        add_generation_prompt=True,
-        **template_kwargs,
-    )
-
-    for key, response in (("chosen", chosen), ("rejected", rejected)):
-        full = tokenizer.apply_chat_template(
-            [dummy_user_message, response],
-            add_generation_prompt=False,
-            **template_kwargs,
-        )
-        result[key] = _extract_response(full, dummy_prompt, response.get("content"))
-
-    return result
+from axolotl.utils.chat_templates import extract_chat_template_args, get_chat_template
+from axolotl.utils.schemas.utils import handle_legacy_message_fields_logic
 
 
 def default(cfg, dataset_idx=0, **kwargs):
@@ -234,8 +50,8 @@ def default(cfg, dataset_idx=0, **kwargs):
         for source in sources:
             role_map[source] = target
 
-    transform_message = _build_message_transform(message_property_mappings, role_map)
-    get_msg_variables = _make_msg_variables_getter()
+    transform_message = build_message_transform(message_property_mappings, role_map)
+    get_msg_variables = make_msg_variables_getter()
 
     def transform_fn(sample, tokenizer=None):
         """Map a dataset sample to prompt/chosen/rejected strings."""
@@ -281,14 +97,14 @@ def default(cfg, dataset_idx=0, **kwargs):
             rejected_msg = rejected_raw[-1]
         rejected = transform_message(rejected_msg, msg_variables)
 
-        return _render_dpo_sample(
+        return render_preference_sample(
             tokenizer,
             messages,
             chosen,
             rejected,
             chat_template_string,
             chat_template_kwargs,
-            _parse_tools(sample.get(field_tools)),
+            parse_tools(sample.get(field_tools)),
         )
 
     return transform_fn, {"remove_columns": [field_messages, field_tools]}
@@ -356,8 +172,8 @@ def argilla_chat(cfg, dataset_idx=0, **kwargs):
         for source in sources:
             role_map[source] = target
 
-    transform_message = _build_message_transform(message_property_mappings, role_map)
-    get_msg_variables = _make_msg_variables_getter()
+    transform_message = build_message_transform(message_property_mappings, role_map)
+    get_msg_variables = make_msg_variables_getter()
 
     def transform_fn(sample, tokenizer=None):
         """Map a dataset sample to prompt/chosen/rejected strings."""
@@ -376,14 +192,14 @@ def argilla_chat(cfg, dataset_idx=0, **kwargs):
         chosen_response = transform_message(chosen_raw[-1], msg_variables)
         rejected_response = transform_message(rejected_raw[-1], msg_variables)
 
-        return _render_dpo_sample(
+        return render_preference_sample(
             tokenizer,
             chosen_messages,
             chosen_response,
             rejected_response,
             chat_template_string,
             chat_template_kwargs,
-            _parse_tools(sample.get(field_tools)),
+            parse_tools(sample.get(field_tools)),
         )
 
     return transform_fn, {"remove_columns": [field_chosen, field_rejected, field_tools]}
